Construct `SourceMap` at the same time as `SessionGlobals`.
Currently `SourceMap` is constructed slightly later than
`SessionGlobals`, and inserted. This commit changes things so they are
done at the same time.

Benefits:
- `SessionGlobals::source_map` changes from
  `Lock<Option<Lrc<SourceMap>>>` to `Option<Lrc<SourceMap>>`. It's still
  optional, but mutability isn't required because it's initialized at
  construction.
- `set_source_map` is removed, simplifying `run_compiler`, which is
  good because that's a critical function and it's nice to make it
  simpler.

This requires moving things around a bit, so the
necessary inputs are available when `SessionGlobals` is created, in
particular the `loader` and `hash_kind`, which are no longer computed by
`build_session`. These inputs are captured by the new `SourceMapInputs`
type, which is threaded through various places. And there's a new