Remove note about iteration count in coerce
and replace it with a simple note suggesting returning a value.

The type mismatch error was never due to how many times
the loop iterates. It is more because of the peculiar structure
of what the for loop desugars to. So the note talking about
iteration count didn't make sense