Remove impl of Allocator for &A
A T-opsem convo on Zulip pointed out that this is hard to optimize-out:

https://rust-lang.zulipchat.com/#narrow/stream/136281-t-opsem/topic/abi.20compat.20of.20.26mut.20T.20and.20*mut.20T.20where.20T.20is.20a.20ZST/near/429718899

This impl merely existing makes the opt story around `Allocated<T, A>`, e.g.
`Vec<T, A>`, `Box<T, A>`, etc. much worse. Now, there may be cases where we
would want this pattern to be used but we can let implementors decide:
We do not need to force everyone to be implicitly opted-into it.