Link rustc/std tools into the correct sysroot
When copying tool binaries, we were linking them into the sysroot of the
compiler that built the binaries. This makes no sense, the binaries are
for the next sysroot. So when the stage0 compiler builds clippy, this
clippy belongs into stage1, and when the stage1 compiler builds clippy,
this clippy belongs into stage2.
This puts it right next to the librustc_driver it actually links
against.
Additionally, we `ensure(Assemble)` of this librustc_driver such that
the tool will be working as expected.
To run the tool manually, we still need to set LD_LIBRARY_PATH, but now
with this, setting the rpath to `$ORIGIN/../lib` (like the `rustc` and
`rustdoc` binaries) should be possible as future work now.

Rustdoc, with its special treatment, was already getting the correct
behavior.