collapse -success and -failure jobs into an -outcome job
Before this commit we had auto-success, auto-failure, try-success and
try-failure. The only difference between a -success and a -failure job
though was the exit code that was returned in the job.

To reduce clutter, this PR collapses each pair of jobs into a single
job that is executed regardless of whether the dependency jobs passed or
failed. Then a simple jq invocation checks whether the dependencies
passed or failed to determine the exit code.