fix: Added missing code example and improved semantics in error explanation text E0283.md
I think there's a missing example in the help text for error E0283. From what I've gathered, it's the same example as the one provided at the beginning of the text so I've added that to where the missing example should be. Also made minor grammatical changes to some sentences to help them flow better.