[bootstrap] Distribute split debuginfo if present
If debuginfo has been requested in `config.toml`, it should be packaged
alongside the appropriate binary when running `x.py dist`.

Currently, this is only implemented for msvc environments where split
debuginfo is (basically) the only option. I've tested that this
correctly packages the `.pdb` for each binary in the various dist
packages.